Create messagePassing.js
Peter Michaux's notes on SICP message passing and how it may be applied in js. 
http://michaux.ca/articles/sicp-classes-for-javascript

In 1) and as in SICP (p.182), makeFromRealImg(x, y) conceptualizes a dispatch procedure that you can call with a message argument. A message is sent to get the real part/magnitude of the complex number.
Example 2) uses the dispatch procedure in a different fashion, calling the procedure instead of directly executing it. 
In 3) speed/performance is improved by taking advantage of javaScript's built in lookup table.